# docs/RequestorAPI.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "requestor-api",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "summary": "Login by Password",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              },
              "example": {
                "email": "admin@example.com",
                "password": "Rahasia123!"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "status_code": 200,
                  "data": {
                    "id": "5cd38571-52a8-4d02-9769-6c08ccd88d9c",
                    "name": "Admin",
                    "email": "admin@example.com",
                    "role": "admin",
                    "status": "active",
                    "created_at": "2026-05-25T11:49:24.575Z",
                    "updated_at": "2026-05-25T11:49:24.575Z",
                    "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjVjZDM4NTcxLTUyYTgtNGQwMi05NzY5LTZjMDhjY2Q4OGQ5YyIsImlhdCI6MTc3OTcxNzMxMiwiZXhwIjoxNzc5ODAzNzEyfQ.4aA_EqVoGy-bjfSoIGoflkEqAnFJK0g-E3eIp7G_UtU",
                    "access_token_expired_at": "2026-05-26T13:55:12.666Z"
                  },
                  "message": "Login By Password successfully"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "summary": "Profile (me)",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "status_code": 200,
                  "data": {
                    "id": "5cd38571-52a8-4d02-9769-6c08ccd88d9c",
                    "name": "Admin",
                    "email": "admin@example.com",
                    "role": "admin",
                    "status": "active",
                    "created_at": "2026-05-25T11:49:24.575Z",
                    "updated_at": "2026-05-25T11:49:24.575Z",
                    "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjVjZDM4NTcxLTUyYTgtNGQwMi05NzY5LTZjMDhjY2Q4OGQ5YyIsImlhdCI6MTc3OTcxNzMxMiwiZXhwIjoxNzc5ODAzNzEyfQ.4aA_EqVoGy-bjfSoIGoflkEqAnFJK0g-E3eIp7G_UtU",
                    "access_token_expired_at": "2026-05-26T13:55:12.666Z"
                  },
                  "message": "Login By Password successfully"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/users/{user_id}": {
      "get": {
        "summary": "Detail User",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "5cd38571-52a8-4d02-9769-6c08ccd88d9c",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "status_code": 200,
                  "data": {
                    "meta": {
                      "total_all_data": 10,
                      "total_view": 10,
                      "max_view": 20,
                      "current_page": 1,
                      "total_page": 1
                    },
                    "items": [
                      {
                        "id": "368d8c23-e707-4736-964c-c3f897e3ec5b",
                        "name": "Viewer 8",
                        "email": "viewer+8@example.com",
                        "role": "viewer",
                        "status": "active",
                        "created_at": "2026-05-25T11:49:24.575Z",
                        "updated_at": "2026-05-25T11:49:24.575Z"
                      },
                      {
                        "id": "12d1ce58-1566-4ec0-8260-dcc0132f704b",
                        "name": "Viewer 7",
                        "email": "viewer+7@example.com",
                        "role": "viewer",
                        "status": "active",
                        "created_at": "2026-05-25T11:49:24.575Z",
                        "updated_at": "2026-05-25T11:49:24.575Z"
                      },
                      {
                        "id": "849bbe15-c745-4f4b-998d-61ddee339aa6",
                        "name": "Viewer 6",
                        "email": "viewer+6@example.com",
                        "role": "viewer",
                        "status": "active",
                        "created_at": "2026-05-25T11:49:24.575Z",
                        "updated_at": "2026-05-25T11:49:24.575Z"
                      },
                      {
                        "id": "92fc512e-e58b-4a1d-aec5-078e8432c256",
                        "name": "Viewer 5",
                        "email": "viewer+5@example.com",
                        "role": "viewer",
                        "status": "active",
                        "created_at": "2026-05-25T11:49:24.575Z",
                        "updated_at": "2026-05-25T11:49:24.575Z"
                      },
                      {
                        "id": "0e8aa680-cd98-464a-96d1-3b0d2701c3b2",
                        "name": "Viewer 4",
                        "email": "viewer+4@example.com",
                        "role": "viewer",
                        "status": "active",
                        "created_at": "2026-05-25T11:49:24.575Z",
                        "updated_at": "2026-05-25T11:49:24.575Z"
                      },
                      {
                        "id": "74cb2127-a6ba-4f0f-a681-bed15a19e9b3",
                        "name": "Viewer 3",
                        "email": "viewer+3@example.com",
                        "role": "viewer",
                        "status": "active",
                        "created_at": "2026-05-25T11:49:24.575Z",
                        "updated_at": "2026-05-25T11:49:24.575Z"
                      },
                      {
                        "id": "c47e5180-3e24-4216-a710-fd40aef416a0",
                        "name": "Viewer 2",
                        "email": "viewer+2@example.com",
                        "role": "viewer",
                        "status": "active",
                        "created_at": "2026-05-25T11:49:24.575Z",
                        "updated_at": "2026-05-25T11:49:24.575Z"
                      },
                      {
                        "id": "5378eaba-d0f8-4e8d-bff1-2984d35f675f",
                        "name": "Viewer 1",
                        "email": "viewer+1@example.com",
                        "role": "viewer",
                        "status": "active",
                        "created_at": "2026-05-25T11:49:24.575Z",
                        "updated_at": "2026-05-25T11:49:24.575Z"
                      },
                      {
                        "id": "40d7da46-da4c-4770-a6ed-e9527dfab74e",
                        "name": "Operator",
                        "email": "operator@example.com",
                        "role": "operator",
                        "status": "active",
                        "created_at": "2026-05-25T11:49:24.575Z",
                        "updated_at": "2026-05-25T11:49:24.575Z"
                      },
                      {
                        "id": "5cd38571-52a8-4d02-9769-6c08ccd88d9c",
                        "name": "Admin",
                        "email": "admin@example.com",
                        "role": "admin",
                        "status": "active",
                        "created_at": "2026-05-25T11:49:24.575Z",
                        "updated_at": "2026-05-25T11:49:24.575Z"
                      }
                    ]
                  },
                  "message": "Users retrieved successfully"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "summary": "Update User by Id",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "5b546735-6d2e-4f36-9384-2a46f1daa18b",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              },
              "example": {
                "name": "testing - updated",
                "email": "testing@gmail.com",
                "password": "Rahasia123!",
                "status": "suspended",
                "role": "viewer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "status_code": 200,
                  "data": {
                    "meta": {
                      "total_all_data": 10,
                      "total_view": 10,
                      "max_view": 20,
                      "current_page": 1,
                      "total_page": 1
                    },
                    "items": [
                      {
                        "id": "368d8c23-e707-4736-964c-c3f897e3ec5b",
                        "name": "Viewer 8",
                        "email": "viewer+8@example.com",
                        "role": "viewer",
                        "status": "active",
                        "created_at": "2026-05-25T11:49:24.575Z",
                        "updated_at": "2026-05-25T11:49:24.575Z"
                      },
                      {
                        "id": "12d1ce58-1566-4ec0-8260-dcc0132f704b",
                        "name": "Viewer 7",
                        "email": "viewer+7@example.com",
                        "role": "viewer",
                        "status": "active",
                        "created_at": "2026-05-25T11:49:24.575Z",
                        "updated_at": "2026-05-25T11:49:24.575Z"
                      },
                      {
                        "id": "849bbe15-c745-4f4b-998d-61ddee339aa6",
                        "name": "Viewer 6",
                        "email": "viewer+6@example.com",
                        "role": "viewer",
                        "status": "active",
                        "created_at": "2026-05-25T11:49:24.575Z",
                        "updated_at": "2026-05-25T11:49:24.575Z"
                      },
                      {
                        "id": "92fc512e-e58b-4a1d-aec5-078e8432c256",
                        "name": "Viewer 5",
                        "email": "viewer+5@example.com",
                        "role": "viewer",
                        "status": "active",
                        "created_at": "2026-05-25T11:49:24.575Z",
                        "updated_at": "2026-05-25T11:49:24.575Z"
                      },
                      {
                        "id": "0e8aa680-cd98-464a-96d1-3b0d2701c3b2",
                        "name": "Viewer 4",
                        "email": "viewer+4@example.com",
                        "role": "viewer",
                        "status": "active",
                        "created_at": "2026-05-25T11:49:24.575Z",
                        "updated_at": "2026-05-25T11:49:24.575Z"
                      },
                      {
                        "id": "74cb2127-a6ba-4f0f-a681-bed15a19e9b3",
                        "name": "Viewer 3",
                        "email": "viewer+3@example.com",
                        "role": "viewer",
                        "status": "active",
                        "created_at": "2026-05-25T11:49:24.575Z",
                        "updated_at": "2026-05-25T11:49:24.575Z"
                      },
                      {
                        "id": "c47e5180-3e24-4216-a710-fd40aef416a0",
                        "name": "Viewer 2",
                        "email": "viewer+2@example.com",
                        "role": "viewer",
                        "status": "active",
                        "created_at": "2026-05-25T11:49:24.575Z",
                        "updated_at": "2026-05-25T11:49:24.575Z"
                      },
                      {
                        "id": "5378eaba-d0f8-4e8d-bff1-2984d35f675f",
                        "name": "Viewer 1",
                        "email": "viewer+1@example.com",
                        "role": "viewer",
                        "status": "active",
                        "created_at": "2026-05-25T11:49:24.575Z",
                        "updated_at": "2026-05-25T11:49:24.575Z"
                      },
                      {
                        "id": "40d7da46-da4c-4770-a6ed-e9527dfab74e",
                        "name": "Operator",
                        "email": "operator@example.com",
                        "role": "operator",
                        "status": "active",
                        "created_at": "2026-05-25T11:49:24.575Z",
                        "updated_at": "2026-05-25T11:49:24.575Z"
                      },
                      {
                        "id": "5cd38571-52a8-4d02-9769-6c08ccd88d9c",
                        "name": "Admin",
                        "email": "admin@example.com",
                        "role": "admin",
                        "status": "active",
                        "created_at": "2026-05-25T11:49:24.575Z",
                        "updated_at": "2026-05-25T11:49:24.575Z"
                      }
                    ]
                  },
                  "message": "Users retrieved successfully"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "summary": "Delete User by Id",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "5b546735-6d2e-4f36-9384-2a46f1daa18b",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local Env"
    },
    {
      "url": "https://requestor-api.dotploy.my.id",
      "description": "Develop Env"
    },
    {
      "url": "https://test.your-api-server.com",
      "description": "Testing Env"
    },
    {
      "url": "https://prod.your-api-server.com",
      "description": "Prod Env"
    }
  ],
  "security": [
    {
      "bearer": []
    }
  ]
}